feat: Add fail case

+++ src/services/CodeService.js
@@ -6,7 +6,15 @@ class CodeService {
     static async createCode(code) {
         await CodeRepo.create(code);
     }
-    static async deleteCode(code_id) {
+    static async deleteCode(code_id, author_id) {
+        (await CodeRepo.findById(code_id)).orElseThrow(new NotFound);
+        const codes_of_author = await CodeRepo.findByAuthorId(author_id);
+        
+        if(codes_of_author.filter(code => code.getId() === code_id).length === 0) {
+            throw new Forbidden;
+        }
+
+
         await CodeRepo.delete(code_id);
     }
     static async modifyCode(code){
